Move idoc partials into /idoc package.
