+++ packages/idoc/src/chapters/001-intro.json
