+++ packages/idoc/src/chapters/002-beginning.json
