+++ packages/idoc/src/essentials/credits.json
