+++ packages/idoc/src/essentials/error.json
