+++ packages/idoc/src/essentials/home.json
