+++ packages/idoc/src/templates/partials/FooBar.js
