--- packages/ui/partials/index.js
@@ -1,4 +1,0 @@
-export { default as FooBar } from './AppBars/FooBar';
-export { default as GlobalStyles } from './Layout/ofLayout/GlobalStyles';
-export { default as Layout } from './Layout/Layout';
-export { default as TopBar } from './AppBars/TopBar';
